+++ zcash_primitives/CHANGELOG.md
@@ -7,10 +7,13 @@ and this library adheres to Rust's notion of
 
 ## [Unreleased]
 ### Added
-- `zcash_primitives::sapling::value`, containing types for handling Sapling note
-  values and value commitments.
+- `zcash_primitives::sapling`:
+  - `keys::{EphemeralSecretKey, EphemeralPublicKey, SharedSecret}`
+  - `value`, containing types for handling Sapling note values and value
+    commitments.
 
 ### Changed
+- MSRV is now 1.60.0.
 - `zcash_primitives::transaction::components::sapling::builder`:
   - `SaplingBuilder::add_output` now takes a
     `zcash_primitives::sapling::value::NoteValue`.